Allow a cross-compilation prefix to be specified
This is really revealing a bad design issue with the build system. The native
build of uKernel should not re-use the object files from the target. This can
only be expected to work when building the i386 or amd64 architectures on a
NetBSD/i386 or NetBSD/amd64 host, with the default toolchain.

More changes are required to fix this.